sustainergy-backend/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
cssselect2==0.6.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-model-utils==4.2.0
django-mysql==4.7.1
djangorestframework==3.13.1
fonttools==4.36.0
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.9.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
pydyf==0.2.0
pygal==3.0.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.11
weasyprint==52.5
webencodings==0.5.1
Werkzeug==2.2.1
xhtml2pdf==0.2.8
zopfli==0.2.1
